+++ lib/src/router.dart
@@ -2,7 +2,7 @@
  * fluro
  * Created by Yakka
  * https://theyakka.com
- * 
+ *
  * Copyright (c) 2018 Yakka, LLC. All rights reserved.
  * See LICENSE for distribution and usage details.
  */
@@ -43,11 +43,13 @@ class Router {
       bool clearStack = false,
       TransitionType transition,
       Duration transitionDuration = const Duration(milliseconds: 250),
-      RouteTransitionsBuilder transitionBuilder}) {
+      RouteTransitionsBuilder transitionBuilder,
+      Object arguments}) {
     RouteMatch routeMatch = matchRoute(context, path,
         transitionType: transition,
         transitionsBuilder: transitionBuilder,
-        transitionDuration: transitionDuration);
+        transitionDuration: transitionDuration,
+        routeSettings: RouteSettings(arguments: arguments));
     Route<dynamic> route = routeMatch.route;
     Completer completer = new Completer();
     Future future = completer.future;
@@ -84,7 +86,8 @@ class Router {
       return new MaterialPageRoute<Null>(
           settings: routeSettings,
           builder: (BuildContext context) {
-            return notFoundHandler.handlerFunc(context, parameters);
+            return notFoundHandler.handlerFunc(
+                context, parameters, routeSettings.arguments);
           });
     };
     return creator(new RouteSettings(name: path), null);
@@ -115,7 +118,7 @@ class Router {
     Map<String, List<String>> parameters =
         match?.parameters ?? <String, List<String>>{};
     if (handler.type == HandlerType.function) {
-      handler.handlerFunc(buildContext, parameters);
+      handler.handlerFunc(buildContext, parameters, routeSettings.arguments);
       return new RouteMatch(matchType: RouteMatchType.nonVisual);
     }
 
@@ -128,7 +131,8 @@ class Router {
             settings: routeSettings,
             fullscreenDialog: transition == TransitionType.nativeModal,
             builder: (BuildContext context) {
-              return handler.handlerFunc(context, parameters);
+              return handler.handlerFunc(
+                  context, parameters, routeSettings.arguments);
             });
       } else {
         var routeTransitionsBuilder;
@@ -141,7 +145,8 @@ class Router {
           settings: routeSettings,
           pageBuilder: (BuildContext context, Animation<double> animation,
               Animation<double> secondaryAnimation) {
-            return handler.handlerFunc(context, parameters);
+            return handler.handlerFunc(
+                context, parameters, routeSettings.arguments);
           },
           transitionDuration: transitionDuration,
           transitionsBuilder: routeTransitionsBuilder,